+++ changes.md
@@ -1,7 +1,10 @@
 #### 0.3.7
 * Added compatibility with Houzez theme.
 * Added compatibility with wp-rabbit plugin.
-* Added `wrc::manage_property::before_update` action to xmlrpc class.
+* Added WP-CLI `wp retsci` commands.
+* Added `wrc::manage_property::before_update` action to XMLRPC class.
+* Added [/wp-json/wp-rets-client/v1/cleanupProcess] endpoint. It recalculates terms counts.
+* Cleaned up XMLRPC class.
 
 #### 0.3.6
 * Added removing of all property attachments on property delete.