+++ changes.md
@@ -1,5 +1,6 @@
 #### 0.3.7
 * Added compatibility with Houzez theme.
+* Added compatibility with Real Homes theme.
 * Added compatibility with wp-rabbit plugin.
 * Added WP-CLI `wp retsci` commands.
 * Added `wrc::manage_property::before_update` action to XMLRPC class.