+++ changes.md
@@ -1,3 +1,6 @@
+#### 0.3.7.1
+* Fixed the bug with replacing fave_agents with default value on updating already existing property.
+
 #### 0.3.7
 * Added compatibility with Houzez theme.
 * Added compatibility with Real Homes theme.